requirements.txt
python-telegram-bot
groq
python-dotenv
psycopg2-binary
pydantic
asyncio
